+++ internal/pkg/valueobject/remote_config.go
@@ -193,7 +193,7 @@ type (
 
 	FaultyPoolsConfig struct {
 		// Min slippage threshold configured in BPS format, ex: 0.01% -> 1, 0.5% -> 50
-		MinSlippageThreshold int64 `mapstructure:"minSlippageThreshold" json:"minSlippageThreshold"`
+		MinSlippageThreshold float64 `mapstructure:"minSlippageThreshold" json:"minSlippageThreshold"`
 	}
 )
